# Optimize dep-graph serialization and improve meta-data hashing