This makes a number of related changes:

- instead of serializing the whole graph, we only serialize `HIR -> WorkProduct` edges
- we gather predecessors more efficiently and cache effort, drastically cutting serialization time by over 6x in my measurements
- compute the metadata hashes more determinstically
- improve `-Z incremental-info`, which tells you how much re-use you are getting, and tells you *why* you are not getting re-use

The end result is that, at least in my tests, we are now able to get 100% re-use when doing `cargo clean; cargo build` twice in a row (with an incremental directory, of course). Here are some rough performance numbers from the `syntex-syntax` crate:

|  build | normal time (debug) | incremental time (debug) |
| ----- | ---- | -------- |
|  full build | 128s | 78s |
|  just syntex_syntax | 96s | 51s |

Note that these numbers only reflect the first phase of incremental, where we are just skipping the trans/LLVM phases. They will get better as we incrementalize more of the compiler. Also, these numbers represent the simplest case, where no actual changes were made. There is still more work to do to handle the case some small number of changes were made correctly (in particular, verifying that the dep-graph is complete etc).

r? @michaelwoerister 

Fixes #35167.