@michaelwoerister 2016-08-06
Are these numbers for a release or for a debug build?