@michaelwoerister 2016-08-08
Ok, I've reviewed everything after `rustfmt save.rs`. The question in `encode_dep_graph` seems important.