@nikomatsakis 2016-08-09
@michaelwoerister OK, I addressed your nits. I will rebase atop of #35166, I think, and maybe try to land them together once you've revised the additional commits in there.